Moved classes to their own files
Also fixed checking for a response type on objects rather than
their class names.
